Display status in with multiple usernames on same realname
Allows the API, cli and web interface to display the status
in case multiple usernames get registered to the same realname.
This introduce an API change int the `client/status` endpoint
which now returns a dict in every cases, otherwise it'll become
difficult for the client handle the data.

Before it the result was:

- no keys: the None string
- one key: a json dict with attributes like {.status } etch
- multiple keys: a json dict with usernames as key and dicts as above
                 as values

Now:

- no keys: empty dict
- multiple / single key: a json dict in the form
                        { "username" : { "status" = "",...}}